requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
seaborn==0.13.2
matplotlib==3.9.2
plotly==5.24.0
scikit-learn==1.5.1
xgboost==2.1.1
scipy==1.13.1
statsmodels==0.14.2
ydata-profiling==4.12.0
sqlglot==25.10.0
chardet==5.2.0
pyyaml==6.0.1
openpyxl==3.1.5
